Migrated trie implementation to separate source
